add detailed example about inner working of the mla reconstruction algorithm
Example 3 shows a step-by-step guide of the different mla-reconstructions steps. the code blocks are set up in a way that they can be copy pasted to different scripts/notebooks with a minimal effort required to make it work. The detailed text block should furthermore provide explanations to the different algorithms used.