implement dc-value offset parsing; dc-value is not included into reconstruction algorithm yet
data_block parsing script is adapted to parse the newline-separated dc-value line corresponding to every datablock. Important to notice:
* the parsing algorithm is not robust against changes in the location of the dc-value line.
* parsing assums that the dc value preceeds the corresponding datablock
* the dc-value is not incuded in the reconstruction yet. this will be done in later changes

Parsing works such that it detets the "dc value:" preceding the data block by string comparison, if detected the dc-data line will be inserted into the daata-block header and the corresponding data-block-parsing function was changed to remove the line again if present. This implementation required multiple if-else clauses at the aforementioned locations in the code.

Other changes:
* implemented a if-hdf5-file-exist-then-remove block before open the hdf5 file for writting to omit the OSError.